[#4] chore: 컴포넌트 재구조화
