+++ src/App.tsx
@@ -9,8 +9,9 @@ import Feedback from "./screens/Feedback";
 import History from "./screens/History";
 import { NavigationContainer } from "@react-navigation/native";
 import { StatusBar } from "expo-status-bar";
-import styled from "styled-components/native";
+import styled, { ThemeProvider } from "styled-components/native";
 import Vote from "./screens/Vote";
+import { theme } from "./styled";
 
 const SafeAreaViewContainer = styled(SafeAreaView)`
   flex: 1;
@@ -29,40 +30,42 @@ const Stack = createNativeStackNavigator<RootStackParamList>();
 
 export default function App() {
   return (
-    <SafeAreaViewContainer>
-      <NavigationContainer>
-        <StatusBar style="dark" />
-        <Stack.Navigator
-          screenOptions={{
-            headerShown: false,
-          }}
-        >
-          <Stack.Screen name="Home" component={Home} />
-          <Stack.Screen name="Matching" component={Matching} />
-          <Stack.Screen
-            name="Chat"
-            component={Chat}
-            options={{
-              headerShown: true,
-              headerBackVisible: false,
-              headerBackTitleVisible: false,
-              title: "", // TODO: Chat Header Hide하는 방법찾기
+    <ThemeProvider theme={theme}>
+      <SafeAreaViewContainer>
+        <NavigationContainer>
+          <StatusBar style="dark" />
+          <Stack.Navigator
+            screenOptions={{
+              headerShown: false,
             }}
-          />
-          <Stack.Screen
-            name="Vote"
-            component={Vote}
-            options={{
-              headerShown: true,
-              headerBackVisible: false,
-              headerBackTitleVisible: false,
-              title: "", // TODO: Chat Header Hide하는 방법찾기
-            }}
-          />
-          <Stack.Screen name="Feedback" component={Feedback} />
-          <Stack.Screen name="History" component={History} />
-        </Stack.Navigator>
-      </NavigationContainer>
-    </SafeAreaViewContainer>
+          >
+            <Stack.Screen name="Home" component={Home} />
+            <Stack.Screen name="Matching" component={Matching} />
+            <Stack.Screen
+              name="Chat"
+              component={Chat}
+              options={{
+                headerShown: true,
+                headerBackVisible: false,
+                headerBackTitleVisible: false,
+                title: "", // TODO: Chat Header Hide하는 방법찾기
+              }}
+            />
+            <Stack.Screen
+              name="Vote"
+              component={Vote}
+              options={{
+                headerShown: true,
+                headerBackVisible: false,
+                headerBackTitleVisible: false,
+                title: "", // TODO: Chat Header Hide하는 방법찾기
+              }}
+            />
+            <Stack.Screen name="Feedback" component={Feedback} />
+            <Stack.Screen name="History" component={History} />
+          </Stack.Navigator>
+        </NavigationContainer>
+      </SafeAreaViewContainer>
+    </ThemeProvider>
   );
 }
