+++ src/constants/color.ts
@@ -1,16 +1,16 @@
 enum Color {
-    PVBlue = "#5988FF",
-    PVSkyblue = "#31BDFF",
-    PVRed = "#EF5D5D",
-    PVYellow = "#FCDE41",
-    PVGreen = "#00AD78",
-    PVBlack01 = "#303134",
-    PVBlack02 = "#6E7081",
-    PVBlack03 = "#AFB1C3",
-    PVBlack04 = "#CBCBD8",
-    PVBlack05 = "#E7E8F0",
-    PVBlack06 = "#F6F6F9",
-    White = "white"
+  PVBlue = "#5988FF",
+  PVSkyblue = "#31BDFF",
+  PVRed = "#EF5D5D",
+  PVYellow = "#FCDE41",
+  PVGreen = "#00AD78",
+  PVBlack01 = "#303134",
+  PVBlack02 = "#6E7081",
+  PVBlack03 = "#AFB1C3",
+  PVBlack04 = "#CBCBD8",
+  PVBlack05 = "#E7E8F0",
+  PVBlack06 = "#F6F6F9",
+  White = "white",
 }
 
-export { Color }
+export { Color };
