+++ src/screens/Chat/ChatUserCardList/ChatUserCardList.tsx
@@ -1,0 +1,16 @@
+import ChatUserCard from "../ChatUserCard";
+import { Container } from "./style";
+import React from "react";
+
+const UserCardList = () => {
+  return (
+    <Container>
+      <ChatUserCard />
+      <ChatUserCard voiceOn={true} />
+      <ChatUserCard />
+      <ChatUserCard />
+    </Container>
+  );
+};
+
+export default UserCardList;
