+++ src/screens/Chat/TooltipFooter/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./TooltipFooter";
