+++ src/screens/Vote/VoteUserCardList/style.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components/native";
+export const Container = styled.View`
+  flex: 1;
+  flex-direction: row;
+  justify-content: flex-start;
+  flex-wrap: wrap;
+`;
