+++ src/screens/Vote/VotingFooter/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./VotingFooter";
