+++ src/screens/Vote/VotingHeader/style.tsx
@@ -1,0 +1,27 @@
+import styled from "styled-components/native";
+
+export const Container = styled.View`
+  flex: 1;
+  flex-direction: row;
+`;
+
+export const MessageSection = styled.View`
+  flex: 2;
+  justify-content: center;
+`;
+
+export const Message = styled.Text`
+  font-weight: 600;
+  font-size: 20px;
+`;
+export const SubMessage = styled.Text`
+  font-weight: 600;
+  font-size: 14px;
+  color: #a1a1ad;
+`;
+
+export const TimerSection = styled.View`
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+`;
