+++ src/styled.ts
@@ -1,0 +1,8 @@
+import { Color } from "./constants/color";
+import { FontSize, FontWeight } from "./constants/font";
+
+export const theme = {
+  color: Color,
+  fontSize: FontSize,
+  fontWeight: FontWeight,
+};
