+++ types/styled.d.ts
@@ -1,0 +1,21 @@
+// src/assets/styles/styled.d.ts
+import "styled-components";
+
+declare module "styled-components" {
+  export interface DefaultTheme {
+    color: {
+      PVBlue: string;
+      PVSkyblue: string;
+      PVRed: string;
+      PVYellow: string;
+      PVGreen: string;
+      PVBlack01: string;
+      PVBlack02: string;
+      PVBlack03: string;
+      PVBlack04: string;
+      PVBlack05: string;
+      PVBlack06: string;
+      White: string;
+    };
+  }
+}
